% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarBoxplot.R, R/radarBoxplot.formula.R
\name{radarBoxplot}
\alias{radarBoxplot}
\alias{radarBoxplot.default}
\alias{radarBoxplot.formula}
\title{Function to plot the Radar-Boxplot}
\usage{
radarBoxplot(x, ...)

\method{radarBoxplot}{default}(x, y, plot.median = F, use.ggplot2 = T,
  mfrow = NA, col = c("red", "blue"), oma = c(5, 4, 0, 0) + 0.1,
  mar = c(0, 0, 1, 1) + 0.1, ...)

\method{radarBoxplot}{formula}(x, data, ...)
}
\arguments{
\item{x}{a data frame or matrix of attributes or a formula describing the
attributes for the class}

\item{...}{optional parameters to be passed to the function `radarBoxplot.default`}

\item{y}{a response vector}

\item{plot.median}{boolean value to flag if median should be plotted, defaults to FALSE}

\item{use.ggplot2}{if ggplot2, data.table and dplyr are available it will use ggplot for plotting. defaults to TRUE}

\item{mfrow}{mfrow argument for defining the subplots nrows and ncols}

\item{col}{the colors to use for radar-boxplot, first color is the percentile 25-75\%
second is the total range and third color is the color for median line}

\item{oma}{outer margins of the subplots}

\item{mar}{margins of the subplots}

\item{data}{optional dataset for which formula was defined}
}
\description{
Function to plot the Radar-Boxplot
}
\examples{
radarBoxplot::radarBoxplot(Species ~ ., iris)

}
