% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winequality_red}
\alias{winequality_red}
\title{Red Wine Quality Dataset}
\format{
A data frame with 1599 rows and 12 variables:
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/wine+quality}
}
\usage{
winequality_red
}
\description{
Related to red vinho verde wine samples, from the
north of Portugal. The goal is to model wine quality
based on physicochemical tests
}
\references{
P. Cortez, A. Cerdeira, F. Almeida, T. Matos and J. Reis.
Modeling wine preferences by data mining from physicochemical properties.
In Decision Support Systems, Elsevier, 47(4):547-553, 2009.
}
\keyword{datasets}
