% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winequality_white}
\alias{winequality_white}
\title{White Wine Quality Dataset}
\format{A data frame with 4898 rows and 12 variables:}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/wine+quality}
}
\usage{
winequality_white
}
\description{
Related to white vinho verde wine samples, from the
north of Portugal. The goal is to model wine quality
based on physicochemical tests
}
\references{
P. Cortez, A. Cerdeira, F. Almeida, T. Matos and J. Reis.
Modeling wine preferences by data mining from physicochemical properties.
In Decision Support Systems, Elsevier, 47(4):547-553, 2009.
}
\keyword{datasets}
