% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarBoxplot.R
\name{radarBoxplot}
\alias{radarBoxplot}
\title{Function to plot the Radar-Boxplot}
\usage{
#Method for formula
radarBoxplot(formula, data, plot.median=FALSE, col=c('red', 'blue'), ...)

#Method for default
radarBoxplot(x, y, plot.median=FALSE, col=c('red', 'blue'), ...)
}
\arguments{
\item{x}{a data frame or matrix of attributes}

\item{formula}{a formula describing the class and attributes to use}

\item{y}{a response vector}

\item{plot.median}{boolean value to flag if median should be plotted, defaults to FALSE}

\item{col}{the colors to use for radar-boxplot, first color is the p25-75%, second is the total range and third color is the color for median line.}
}
\description{
Plots the Radar-Boxplot
}
\examples{
data(iris)
radarBoxplot(Species ~ ., iris)
}
\keyword{boxplot}
\keyword{radar}
